Support TestClient_ExecuteProgram on Windows
