+++ internal/runnerv2client/client_test.go
@@ -3,7 +3,6 @@ package runnerv2client
 import (
 	"bytes"
 	"context"
-	"io"
 	"net"
 	"testing"
 	"time"
@@ -102,41 +101,6 @@ func TestClient_ExecuteProgram(t *testing.T) {
 		)
 		require.NoError(t, err)
 		require.Equal(t, "test-input-non-interactive\n", stdout.String())
-	})
-
-	t.Run("InputInteractive", func(t *testing.T) {
-		t.Parallel()
-
-		client := testCreateClient(t, lis)
-		ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
-		defer cancel()
-
-		cfg := &command.ProgramConfig{
-			ProgramName: "bash",
-			Source: &runnerv2.ProgramConfig_Commands{
-				Commands: &runnerv2.ProgramConfig_CommandList{
-					Items: []string{
-						"read -r name",
-						"echo $name",
-					},
-				},
-			},
-			Interactive: true,
-			Mode:        runnerv2.CommandMode_COMMAND_MODE_INLINE,
-		}
-		stdout := new(bytes.Buffer)
-		err := client.ExecuteProgram(
-			ctx,
-			cfg,
-			ExecuteProgramOptions{
-				Stdin:  io.NopCloser(bytes.NewBufferString("test-input-interactive\n")),
-				Stdout: stdout,
-			},
-		)
-		require.NoError(t, err)
-		// Using [require.Contains] because on Linux the input is repeated.
-		// Unclear why it passes fine on macOS.
-		require.Contains(t, stdout.String(), "test-input-interactive\r\n")
 	})
 }
